Pankow Kategorien. Not complete. A big mess :-(
Pankow brings a big mess-up :-(
A lot of categories which exist already as "Angebot" in other Bezirke, are re-introduced here as "Kategorie".
The "Kategorie" entries are now bloated.

This is another indication that "Kategorie" is a very bad criteria name.
It seems it was never decided what the exact difference between "Kategorie" and "Angebot" is.

My urgent suggestion:
- rename "Kategorie" to "Thema".
- consider "Thema" as an abstract generic term to group the actual "Angebote". Thema are e.g. "Bildung", "Familie", "Sport", "Gesundheit", "Migranten", "Wohnen", "Umwelt".
- "Angebote" are the actual things where you can go to or participate, e.g. "Schulen", "Kindertagesstätten", "Familienberatung", "Spielplätze", "Treffpunkte", "Schuldnerberatung", "Ärzte".